Added centralized api router and requestor classes
The CSResponsiveApiRouter class holds the routes to the api and the way
to make requests on it.  It contains an instance of the Requestor
protocol to make the actual requests.  This class can be changed based
on the current environment to hit different ip addresses or not hit the
api at all.

A Faked Api Response class was added that returns 2 feeds or 10 feed
items based on which response is called.  It does not currently take
parameters to change the output.  The feed create action is not yet
supported.

The remote object base class was added to the models folder to handle
the object mapping.  A naive implementation of the object mapping and
saving was added to the success callback generator method in the
ApiRouter.  The externalID fields on the Feed and FeedItem models were
changed to id to make the object mapping simpler.  The inflections pod,
nil_additions category, and the property_reflection category were added
to make the object mapper work.

[Finishes #66970134]
[Finishes #66953872]